maachbazar-py/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
pyngrok
supabase
